rename breeding statistics to breeding scheme
